Go back to top of document if no hash
I found trying to navigate the docs difficult because I'd be in the table of contents at the top of an API doc. I'd click a function then click back but it didn't go back to where I was. This not perfect since it not technically "where I was" but it's closer than not going anywhere which is what it does now.

wdyt?